Add support for test framework Vitest, extend README and update dependencies
Add support for the testing framework Vitest has been added in vite.config.ts and a basic Vitest unit test has been developed in App.test.tsx.

The addition of Vitest allows for testing of our application in a more straightforward and efficient way.

README was also expanded to provide more information on the project purpose and goals.